rename scripts and update documentation
